Process mouse events, including button presses and movement.

+++ include/BaseTypes.h
@@ -2,16 +2,17 @@
 #define BASETPYES_H
 
 // platform types for windows 64 bit
+#include <cstdint>
 
-typedef unsigned char            T_UINT8;
-typedef unsigned short int       T_UINT16;
-typedef unsigned long int        T_UINT32;
-typedef unsigned long long int   T_UINT64;
-typedef signed char              T_INT8;
-typedef signed short int         T_INT16;
-typedef signed long int          T_INT32;
-typedef signed long long int     T_INT64;
-typedef float                    T_FLOAT32;
-typedef double                   T_FLOAT64;
+typedef unsigned char       T_UINT8;
+typedef unsigned short      T_UINT16;
+typedef unsigned int        T_UINT32;
+typedef unsigned long long  T_UINT64;
+typedef char                T_INT8;
+typedef short               T_INT16;
+typedef int                 T_INT32;
+typedef long long           T_INT64;
+typedef float               T_FLOAT32;
+typedef double              T_FLOAT64;
 
 #endif
